Move stop/start system to extenal file so it could easily replaced
