+++ tests/base_test_agent.py
@@ -5,6 +5,7 @@ import datetime
 from subprocess import call
 import spur
 import os
+import dockers
 
 class BaseDevnetAgent:
     def __init__(self):
